Dave Perry:
I have modeled N7764P, the comanche 250 I co-own with 2 other Seagate engineers.

The panel layout is from photos of the actual panel.  I used an edit of a
digital photo to make the texture for the fuel selector and since there is only
one fuel quantity gage that is connected to the selected tank, I modeled this
as well.  Had my first try at using nasal to make this work.  I was quite
surprised that yasim would allow me to get the clean stall correct, the flaps
down, gear down stall correct, and the cruise speeds correct at various
altitudes.  The one area that I want to improve with help from experienced
yasim people is the match of low manifold pressure performance.