Update from Dave Perry:
I removed the local controls.nas and added the flapsDown and gearDown in
pa24.nas copying/editing the b29 flapsDown function that uses
interpolation to achieve the continuous motion.  This allows all the
flap and gear inputs to use the standard named functions.  So I removed
the local bindings to "[" and "]".

Also, I made the fastpropDisc smoothly vanish as rpm increases.  I
really like this effect!