Changes from Dave Perry:
Melchior sent me a note off-list with a minor bug in pa24.nas.  So I
combined that fix with an adjustment of the right aileron offset and
moved the pilot into the main .ac model so that the pilot and seat are
in the AC in the startup of fgrun (looks more complete).

1. Either add face.rgb to Models, or move face.rgb from Models/pilot to
Models.
2. Remove the folder Models/pilot
3. Apply the patch